After merging CompiledKernel updates
