+++ python/runtime/cudaq/platform/JITExecutionCache.cpp
@@ -7,7 +7,6 @@
  ******************************************************************************/
 #include "JITExecutionCache.h"
 
-using namespace cudaq_internal::compiler;
 using namespace mlir;
 
 namespace cudaq {
