+++ runtime/common/ServerHelper.h
@@ -8,6 +8,7 @@
 
 #pragma once
 
+#include "CompiledKernel.h"
 #include "ExecutionContext.h"
 #include "Future.h"
 #include "Registry.h"
@@ -15,7 +16,6 @@
 #include "RuntimeTarget.h"
 #include "SampleResult.h"
 #include "common/RecordLogParser.h"
-#include "cudaq_internal/compiler/JIT.h"
 #include "nlohmann/json.hpp"
 #include <filesystem>
 
@@ -31,19 +31,19 @@ using BackendConfig = std::map<std::string, std::string>;
 struct KernelExecution {
   std::string name;
   std::string code;
-  std::optional<cudaq_internal::compiler::JitEngine> jit;
+  std::optional<cudaq::JitEngine> jit;
   std::optional<Resources> resourceCounts;
   nlohmann::json output_names;
   std::vector<std::size_t> mapping_reorder_idx;
   nlohmann::json user_data;
   KernelExecution(std::string &n, std::string &c,
-                  std::optional<cudaq_internal::compiler::JitEngine> jit,
+                  std::optional<cudaq::JitEngine> jit,
                   std::optional<Resources> rc, nlohmann::json &o,
                   std::vector<std::size_t> &m)
       : name(n), code(c), jit(jit), resourceCounts(rc), output_names(o),
         mapping_reorder_idx(m) {}
   KernelExecution(std::string &n, std::string &c,
-                  std::optional<cudaq_internal::compiler::JitEngine> jit,
+                  std::optional<cudaq::JitEngine> jit,
                   std::optional<Resources> rc, nlohmann::json &o,
                   std::vector<std::size_t> &m, nlohmann::json &ud)
       : name(n), code(c), jit(jit), resourceCounts(rc), output_names(o),
